requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pygetwindow>=0.0.9
pyautogui>=0.9.54
keyboard>=0.13.5
matplotlib>=3.7.0
psutil>=5.9.0
